+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/parameters/PayloadAttributesV4.java
@@ -12,39 +12,32 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.engine;
+package org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters;
 
-import org.hyperledger.besu.datatypes.Address;
+import org.hyperledger.besu.ethereum.core.Withdrawal;
+
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.apache.tuweni.bytes.Bytes32;
 
-public sealed class PayloadAttributesV1 permits PayloadAttributesV2 {
+public final class PayloadAttributesV4 extends PayloadAttributesV3 {
 
-  private final long timestamp;
-  private final Bytes32 prevRandao;
-  private final Address suggestedFeeRecipient;
+  private final Long slotNumber;
 
   @JsonCreator
-  public PayloadAttributesV1(
+  public PayloadAttributesV4(
       @JsonProperty("timestamp") final String timestamp,
       @JsonProperty("prevRandao") final String prevRandao,
-      @JsonProperty("suggestedFeeRecipient") final String suggestedFeeRecipient) {
-    this.timestamp = Long.decode(timestamp);
-    this.prevRandao = Bytes32.fromHexString(prevRandao);
-    this.suggestedFeeRecipient = Address.fromHexString(suggestedFeeRecipient);
+      @JsonProperty("suggestedFeeRecipient") final String suggestedFeeRecipient,
+      @JsonProperty("withdrawals") final List<Withdrawal> withdrawals,
+      @JsonProperty("parentBeaconBlockRoot") final String parentBeaconBlockRoot,
+      @JsonProperty("slotNumber") final String slotNumber) {
+    super(timestamp, prevRandao, suggestedFeeRecipient, withdrawals, parentBeaconBlockRoot);
+    this.slotNumber = slotNumber != null ? Long.decode(slotNumber) : null;
   }
 
-  public long getTimestamp() {
-    return timestamp;
-  }
-
-  public Bytes32 getPrevRandao() {
-    return prevRandao;
-  }
-
-  public Address getSuggestedFeeRecipient() {
-    return suggestedFeeRecipient;
+  public Long getSlotNumber() {
+    return slotNumber;
   }
 }
