+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineForkchoiceUpdatedV4Test.java
@@ -12,11 +12,10 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.engine.forkchoiceupdated;
+package org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.engine;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hyperledger.besu.datatypes.HardforkId.MainnetHardforkId.AMSTERDAM;
-import static org.hyperledger.besu.datatypes.HardforkId.MainnetHardforkId.CANCUN;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -26,8 +25,8 @@ import org.hyperledger.besu.consensus.merge.blockcreation.PayloadIdentifier;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.engine.ForkchoiceStateV1;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.engine.PayloadAttributesV3;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.ForkchoiceStateV1;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.PayloadAttributesV4;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponse;
@@ -46,78 +45,82 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-public class EngineForkchoiceUpdatedV3Test extends EngineForkchoiceUpdatedV2Test {
-
-  static final long AMSTERDAM_MILESTONE = 2_000_000L;
+public class EngineForkchoiceUpdatedV4Test extends EngineForkchoiceUpdatedV3Test {
 
   @Override
   protected EngineForkchoiceUpdatedV1<?> createMethodInstance() {
-    return new EngineForkchoiceUpdatedV3<>(
+    // V4 has no upper bound (null maxFork = open-ended).
+    return new EngineForkchoiceUpdatedV4<>(
         vertx,
         protocolSchedule,
         protocolContext,
         mergeCoordinator,
         engineCallListener,
-        CANCUN,
-        AMSTERDAM);
+        AMSTERDAM,
+        null);
   }
 
   @Override
   @BeforeEach
   public void before() {
     super.before();
-    // AMSTERDAM upper bound; set blockHeaderBuilder default timestamp to CANCUN_MILESTONE so
-    // inherited tests that call buildHeader() produce payloads in the valid [CANCUN, AMSTERDAM)
-    // window.
-    when(protocolSchedule.milestoneFor(AMSTERDAM)).thenReturn(Optional.of(AMSTERDAM_MILESTONE));
-    blockHeaderBuilder.timestamp(CANCUN_MILESTONE);
+    // Set blockHeaderBuilder default timestamp to AMSTERDAM_MILESTONE so inherited tests that call
+    // blockHeaderBuilder.buildHeader() produce payloads >= AMSTERDAM_MILESTONE (V4's lower bound).
+    blockHeaderBuilder.timestamp(AMSTERDAM_MILESTONE);
     createMethod();
   }
 
   @Override
   @Test
   public void shouldReturnExpectedMethodName() {
-    assertThat(method.getName()).isEqualTo("engine_forkchoiceUpdatedV3");
+    assertThat(method.getName()).isEqualTo("engine_forkchoiceUpdatedV4");
   }
 
   @Override
   protected String getMethodName() {
-    return RpcMethod.ENGINE_FORKCHOICE_UPDATED_V3.getMethodName();
+    return RpcMethod.ENGINE_FORKCHOICE_UPDATED_V4.getMethodName();
   }
 
   @Override
   protected Object validPayloadAttributesForBlock(final BlockHeader head) {
-    return new PayloadAttributesV3(
+    return new PayloadAttributesV4(
         String.valueOf(head.getTimestamp() + 1),
         Bytes32.fromHexStringLenient("0xDEADBEEF").toHexString(),
         Address.ECREC.toString(),
         Collections.emptyList(),
-        Bytes32.ZERO.toHexString());
+        Bytes32.ZERO.toHexString(),
+        "0x1");
   }
 
   @Override
   protected Object invalidTimestampPayloadAttributesForBlock(final BlockHeader head) {
-    return new PayloadAttributesV3(
+    return new PayloadAttributesV4(
         String.valueOf(head.getTimestamp()),
         Bytes32.fromHexStringLenient("0xDEADBEEF").toHexString(),
         Address.ECREC.toString(),
         Collections.emptyList(),
-        Bytes32.ZERO.toHexString());
+        Bytes32.ZERO.toHexString(),
+        "0x1");
   }
 
-  // Inherited from V2Test: in V3, Cancun-era timestamps are within the valid [CANCUN, AMSTERDAM)
-  // window — the V2 upper-bound scenario no longer applies here.
+  // Inherited from V3Test: V4 has no upper bound, so Amsterdam timestamps are valid (open-ended).
   @Override
   @Test
-  public void shouldReturnUnsupportedForkIfBlockTimestampIsAfterCancunMilestone() {}
+  public void shouldReturnUnsupportedForkIfBlockTimestampIsAfterAmsterdamMilestone() {}
 
-  // ---- V3-specific tests ----
+  // Inherited from V3Test: Cancun timestamps (< Amsterdam) are below V4's minimum — not applicable.
+  @Override
+  @Test
+  public void shouldReturnValidForTimestampInCancunWindow() {}
+
+  // ---- V4-specific tests ----
 
   @Test
-  public void shouldReturnUnsupportedForkIfBlockTimestampIsBeforeCancunMilestone() {
-    // head.timestamp = CANCUN_MILESTONE-2 so payload = CANCUN_MILESTONE-1 (> head) and
-    // CANCUN_MILESTONE-1 < CANCUN_MILESTONE triggers the min-bound UNSUPPORTED_FORK
-    final BlockHeader mockHeader = blockHeaderBuilder.timestamp(CANCUN_MILESTONE - 2).buildHeader();
+  public void shouldReturnUnsupportedForkIfBlockTimestampIsBeforeAmsterdamMilestone() {
+    // head.timestamp = AMSTERDAM_MILESTONE-2 so payload = AMSTERDAM_MILESTONE-1 (> head) and
+    // AMSTERDAM_MILESTONE-1 < AMSTERDAM_MILESTONE triggers the min-bound UNSUPPORTED_FORK
+    final BlockHeader mockHeader =
+        blockHeaderBuilder.timestamp(AMSTERDAM_MILESTONE - 2).buildHeader();
     when(mergeCoordinator.getOrSyncHeadByHash(mockHeader.getHash(), Hash.ZERO))
         .thenReturn(Optional.of(mockHeader));
     when(mergeCoordinator.isDescendantOf(any(), any())).thenReturn(true);
@@ -139,33 +142,8 @@ public class EngineForkchoiceUpdatedV3Test extends EngineForkchoiceUpdatedV2Test
   }
 
   @Test
-  public void shouldReturnUnsupportedForkIfBlockTimestampIsAfterAmsterdamMilestone() {
-    // head.timestamp = AMSTERDAM_MILESTONE so payload = AMSTERDAM_MILESTONE+1 (> head) and
-    // AMSTERDAM_MILESTONE+1 >= AMSTERDAM_MILESTONE triggers the max-bound UNSUPPORTED_FORK
+  public void shouldReturnValidForTimestampAtAmsterdamMilestone() {
     final BlockHeader mockHeader = blockHeaderBuilder.timestamp(AMSTERDAM_MILESTONE).buildHeader();
-    when(mergeCoordinator.getOrSyncHeadByHash(mockHeader.getHash(), Hash.ZERO))
-        .thenReturn(Optional.of(mockHeader));
-    when(mergeCoordinator.isDescendantOf(any(), any())).thenReturn(true);
-    when(mergeCoordinator.computeReorgDepth(any())).thenReturn(OptionalLong.empty());
-    when(mergeCoordinator.updateForkChoiceWithoutLegacySkip(any(), any(), any()))
-        .thenReturn(mock(ForkchoiceResult.class));
-
-    final JsonRpcResponse resp =
-        resp(
-            new ForkchoiceStateV1(mockHeader.getBlockHash(), Hash.ZERO, Hash.ZERO),
-            Optional.of(validPayloadAttributesForBlock(mockHeader)));
-
-    final JsonRpcError jsonRpcError =
-        Optional.of(resp)
-            .map(JsonRpcErrorResponse.class::cast)
-            .map(JsonRpcErrorResponse::getError)
-            .get();
-    assertThat(jsonRpcError.getCode()).isEqualTo(RpcErrorType.UNSUPPORTED_FORK.getCode());
-  }
-
-  @Test
-  public void shouldReturnValidForTimestampInCancunWindow() {
-    final BlockHeader mockHeader = blockHeaderBuilder.timestamp(CANCUN_MILESTONE).buildHeader();
     when(mergeCoordinator.getOrSyncHeadByHash(mockHeader.getHash(), Hash.ZERO))
         .thenReturn(Optional.of(mockHeader));
     when(mergeCoordinator.isDescendantOf(any(), any())).thenReturn(true);
@@ -178,6 +156,65 @@ public class EngineForkchoiceUpdatedV3Test extends EngineForkchoiceUpdatedV2Test
         resp(
             new ForkchoiceStateV1(mockHeader.getBlockHash(), Hash.ZERO, Hash.ZERO),
             Optional.of(validPayloadAttributesForBlock(mockHeader)));
+
+    assertThat(resp).isInstanceOf(JsonRpcSuccessResponse.class);
+  }
+
+  @Test
+  public void shouldReturnInvalidSlotNumberParamsForNegativeSlotNumber() {
+    final BlockHeader mockHeader =
+        blockHeaderBuilder.timestamp(AMSTERDAM_MILESTONE + 1).buildHeader();
+    when(mergeCoordinator.getOrSyncHeadByHash(mockHeader.getHash(), Hash.ZERO))
+        .thenReturn(Optional.of(mockHeader));
+    when(mergeCoordinator.isDescendantOf(any(), any())).thenReturn(true);
+    when(mergeCoordinator.computeReorgDepth(any())).thenReturn(OptionalLong.empty());
+    when(mergeCoordinator.updateForkChoiceWithoutLegacySkip(any(), any(), any()))
+        .thenReturn(mock(ForkchoiceResult.class));
+
+    final PayloadAttributesV4 payloadWithNegativeSlot =
+        new PayloadAttributesV4(
+            String.valueOf(mockHeader.getTimestamp() + 1),
+            Bytes32.fromHexStringLenient("0xDEADBEEF").toHexString(),
+            Address.ECREC.toString(),
+            Collections.emptyList(),
+            Bytes32.ZERO.toHexString(),
+            "-0x1");
+
+    final JsonRpcResponse resp =
+        resp(
+            new ForkchoiceStateV1(mockHeader.getBlockHash(), Hash.ZERO, Hash.ZERO),
+            Optional.of(payloadWithNegativeSlot));
+
+    assertThat(resp).isInstanceOf(JsonRpcErrorResponse.class);
+    assertThat(((JsonRpcErrorResponse) resp).getErrorType())
+        .isEqualTo(RpcErrorType.INVALID_SLOT_NUMBER_PARAMS);
+  }
+
+  @Test
+  public void shouldReturnValidForZeroSlotNumber() {
+    final BlockHeader mockHeader =
+        blockHeaderBuilder.timestamp(AMSTERDAM_MILESTONE + 1).buildHeader();
+    when(mergeCoordinator.getOrSyncHeadByHash(mockHeader.getHash(), Hash.ZERO))
+        .thenReturn(Optional.of(mockHeader));
+    when(mergeCoordinator.isDescendantOf(any(), any())).thenReturn(true);
+    when(mergeCoordinator.computeReorgDepth(any())).thenReturn(OptionalLong.empty());
+    when(mergeCoordinator.preparePayload(any())).thenReturn(new PayloadIdentifier(1337L));
+    when(mergeCoordinator.updateForkChoiceWithoutLegacySkip(any(), any(), any()))
+        .thenReturn(ForkchoiceResult.withResult(Optional.empty(), Optional.of(mockHeader)));
+
+    final PayloadAttributesV4 payloadWithZeroSlot =
+        new PayloadAttributesV4(
+            String.valueOf(mockHeader.getTimestamp() + 1),
+            Bytes32.fromHexStringLenient("0xDEADBEEF").toHexString(),
+            Address.ECREC.toString(),
+            Collections.emptyList(),
+            Bytes32.ZERO.toHexString(),
+            "0x0");
+
+    final JsonRpcResponse resp =
+        resp(
+            new ForkchoiceStateV1(mockHeader.getBlockHash(), Hash.ZERO, Hash.ZERO),
+            Optional.of(payloadWithZeroSlot));
 
     assertThat(resp).isInstanceOf(JsonRpcSuccessResponse.class);
   }
